baseinput: updated css + added documentation
